Add files via upload
This project focuses on improving the quality of images using OpenCV. It works by first reducing noise to remove any unwanted spots or imperfections in the image. Then, it applies filters to smooth out the picture while keeping important details, like edges, sharp and clear. After that, it enhances the sharpness to make the image look more defined and polished. The goal is to take an image, clean it up, and make it look clearer and more professional by using a combination of filtering and sharpening techniques.